Update the touch mode so that the last element is not focused Enact-DCO-1.0-Signed-off-by: SJ RO (sj.ro@lge.com)
